pyproject.toml
[tool.poetry]
name = "private-assistant-skill-coordinator"
version = "0.1.0"
description = "Coordinator for Modular Private Assistant: Manages skill-response certainties and orchestrates message processing via MQTT."
authors = ["stkr22 <stkr22@github.com>"]
readme = "README.md"
license = "GNU GENERAL PUBLIC LICENSE"
include = ["src"]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
paho-mqtt = "^2.0.0"
pyyaml = "^6.0.1"
pydantic = "^2.7.0"
typer = {extras = ["all"], version = "^0.12.3"}

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.11"
mypy = "^1.8.0"
pytest = "^7.4.4"
types-pyyaml = "^6.0.12.20240311"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "UP"]
ignore = ["E203"]
